Refactor ead search overview into two queries.
When you separate the aggregation and the collapse query the
ElasticSearch caches are used more efficiently. The result is a
significantly faster query and less load on the elasticsearch cluster.
